Move commands.InputHandler to commands/input.Handler
The InputHandler was getting complicated enough to need multiple files, so it has been moved to its own package.
